[cfgmgr]: Integrate MACsecMGR with sonic-buildimage (#1627)
1. Fix bug about the return value of `get_value`, it should return `true` if it's OK
2. Fix bug about the type of `MACsecProfile::priority`, the original `uint8_t` will cause a `lexical_convert` failure when the priority is a integer.
3. Polish log in `get_value`.
4. Change the predefined paths for sonic-buildimage integration

+++ cfgmgr/macsecmgr.h
@@ -27,7 +27,7 @@ public:
     using TaskArgs = std::vector<FieldValueTuple>;
     struct MACsecProfile
     {
-        std::uint8_t        priority;
+        std::uint32_t       priority;
         std::string         cipher_suite;
         std::string         primary_cak;
         std::string         primary_ckn;
